Split Android app into MVVM structure
