+++ app/lib/src/l10n/app_strings.dart
@@ -1,0 +1,400 @@
+part of '../../main.dart';
+
+class AppStrings {
+  const AppStrings(this.locale);
+
+  final Locale locale;
+
+  static const supportedLocales = [
+    Locale('en'),
+    Locale('ja'),
+    Locale('zh'),
+    Locale('ko'),
+  ];
+
+  static const LocalizationsDelegate<AppStrings> delegate =
+      _AppStringsDelegate();
+
+  static AppStrings of(BuildContext context) =>
+      Localizations.of<AppStrings>(context, AppStrings)!;
+
+  String get code {
+    final languageCode = locale.languageCode.toLowerCase();
+    if (_messages.containsKey(languageCode)) {
+      return languageCode;
+    }
+    return 'en';
+  }
+
+  String t(String key) => _messages[code]?[key] ?? _messages['en']![key] ?? key;
+}
+
+class _AppStringsDelegate extends LocalizationsDelegate<AppStrings> {
+  const _AppStringsDelegate();
+
+  @override
+  bool isSupported(Locale locale) => AppStrings.supportedLocales.any(
+    (item) => item.languageCode == locale.languageCode,
+  );
+
+  @override
+  Future<AppStrings> load(Locale locale) async => AppStrings(locale);
+
+  @override
+  bool shouldReload(covariant LocalizationsDelegate<AppStrings> old) => false;
+}
+
+extension AppStringsContext on BuildContext {
+  AppStrings get l10n => AppStrings.of(this);
+}
+
+const _messages = <String, Map<String, String>>{
+  'en': {
+    'signingIn': 'Signing in',
+    'preparingRelay': 'Preparing secure relay access.',
+    'startupFailed': 'Startup failed',
+    'sessionLoadFailed': 'Session load failed',
+    'commandLoadFailed': 'Command load failed',
+    'newSession': 'New session',
+    'create': 'Create',
+    'creating': 'Creating',
+    'connectedAnonymous': 'Connected as anonymous user',
+    'pcBridge': 'PC bridge',
+    'lastHeartbeat': 'Last heartbeat',
+    'lastQueueCheck': 'Last queue check',
+    'lastManualCheck': 'Last manual check',
+    'lastResponse': 'Last response',
+    'checkPcNow': 'Check PC now',
+    'checking': 'Checking',
+    'cliDefaults': 'CLI defaults',
+    'loading': 'Loading',
+    'notifications': 'Notifications',
+    'settings': 'Settings',
+    'connectionSettings': 'Connection settings',
+    'notSeenYet': 'Not seen yet',
+    'sessions': 'Sessions',
+    'sessionCount': 'session(s)',
+    'loadingSessions': 'Loading sessions...',
+    'noSessionsYet': 'No sessions yet',
+    'noMatchingSessions': 'No matching sessions',
+    'noCommandsYet': 'No commands yet',
+    'searchSessions': 'Search sessions',
+    'allGroups': 'All groups',
+    'ungrouped': 'Ungrouped',
+    'favorites': 'Favorites',
+    'favorite': 'Favorite',
+    'removeFavorite': 'Remove favorite',
+    'renameSession': 'Rename session',
+    'sessionName': 'Session name',
+    'group': 'Group',
+    'groupName': 'Group name',
+    'changeGroup': 'Change group',
+    'deleteSession': 'Delete session',
+    'deleteSessionQuestion': 'Delete this session from the app history?',
+    'delete': 'Delete',
+    'more': 'More',
+    'waitingFinalResult': 'Waiting for final result.',
+    'elapsed': 'Elapsed',
+    'lastProgress': 'Last progress',
+    'instruction': 'Instruction',
+    'send': 'Send',
+    'model': 'Model',
+    'profile': 'Profile',
+    'sandbox': 'Sandbox',
+    'bypassSandbox': 'Bypass sandbox',
+    'bypassSandboxSubtitle': 'Overrides the sandbox selection',
+    'sandboxUsesDefaults': 'Sandbox and bypass use CLI defaults.',
+    'advancedCliOptions': 'Advanced CLI options',
+    'optionalConfigProfile': 'Optional config profile',
+    'defaultOption': 'Default',
+    'bridgeReadsFinalOutput': 'Bridge still reads final output from file',
+    'help': 'Help',
+    'cancel': 'Cancel',
+    'save': 'Save',
+    'close': 'Close',
+    'selectImageFile': 'Select image file',
+    'showHelpFor': 'Show help for',
+    'cliOptionHelp': 'CLI option help',
+    'noHelpAvailable': 'No help is available for this option.',
+    'where': 'Where',
+    'example': 'Example',
+    'none': 'None',
+    'on': 'on',
+    'off': 'off',
+    'config': 'Config',
+    'enable': 'Enable',
+    'disable': 'Disable',
+    'images': 'Images',
+    'oss': 'OSS',
+    'localProvider': 'Local provider',
+    'fullAuto': 'Full auto',
+    'addDirs': 'Add dirs',
+    'skipGitRepoCheck': 'Skip git repo check',
+    'ephemeral': 'Ephemeral',
+    'ignoreUserConfig': 'Ignore user config',
+    'ignoreRules': 'Ignore rules',
+    'outputSchema': 'Output schema',
+    'jsonEvents': 'JSON events',
+  },
+  'ja': {
+    'signingIn': 'サインイン中',
+    'preparingRelay': '安全なリレー接続を準備しています。',
+    'startupFailed': '起動に失敗しました',
+    'sessionLoadFailed': 'セッションの読み込みに失敗しました',
+    'commandLoadFailed': 'コマンドの読み込みに失敗しました',
+    'newSession': '新規セッション',
+    'create': '作成',
+    'creating': '作成中',
+    'connectedAnonymous': '匿名ユーザーで接続中',
+    'pcBridge': 'PCブリッジ',
+    'lastHeartbeat': '最終heartbeat',
+    'lastQueueCheck': '最終queue確認',
+    'lastManualCheck': '最終手動確認',
+    'lastResponse': '最終応答',
+    'checkPcNow': 'PCを確認',
+    'checking': '確認中',
+    'cliDefaults': 'CLI既定値',
+    'loading': '読込中',
+    'notifications': '通知',
+    'settings': '設定',
+    'connectionSettings': '接続設定',
+    'notSeenYet': '未確認',
+    'sessions': 'セッション',
+    'sessionCount': '件',
+    'loadingSessions': 'セッションを読み込み中...',
+    'noSessionsYet': 'セッションはまだありません',
+    'noMatchingSessions': '一致するセッションはありません',
+    'noCommandsYet': 'コマンドはまだありません',
+    'searchSessions': 'セッションを検索',
+    'allGroups': 'すべてのグループ',
+    'ungrouped': '未分類',
+    'favorites': 'お気に入り',
+    'favorite': 'お気に入り',
+    'removeFavorite': 'お気に入りを解除',
+    'renameSession': 'セッション名を変更',
+    'sessionName': 'セッション名',
+    'group': 'グループ',
+    'groupName': 'グループ名',
+    'changeGroup': 'グループを変更',
+    'deleteSession': 'セッションを削除',
+    'deleteSessionQuestion': 'このセッションをアプリの履歴から削除しますか？',
+    'delete': '削除',
+    'more': 'その他',
+    'waitingFinalResult': '最終結果を待っています。',
+    'elapsed': '経過時間',
+    'lastProgress': '最終進捗',
+    'instruction': '指示',
+    'send': '送信',
+    'model': 'モデル',
+    'profile': 'プロファイル',
+    'sandbox': 'サンドボックス',
+    'bypassSandbox': 'サンドボックス迂回',
+    'bypassSandboxSubtitle': 'サンドボックス選択より優先します',
+    'sandboxUsesDefaults': 'サンドボックスと迂回設定はCLI既定値を使用します。',
+    'advancedCliOptions': '詳細CLIオプション',
+    'optionalConfigProfile': '任意のconfig profile',
+    'defaultOption': '既定',
+    'bridgeReadsFinalOutput': '最終出力は引き続きファイルから読み取ります',
+    'help': 'ヘルプ',
+    'cancel': 'キャンセル',
+    'save': '保存',
+    'close': '閉じる',
+    'selectImageFile': '画像ファイルを選択',
+    'showHelpFor': 'ヘルプを表示',
+    'cliOptionHelp': 'CLIオプションヘルプ',
+    'noHelpAvailable': 'この項目のヘルプはありません。',
+    'where': '表示場所',
+    'example': '入力例',
+    'none': 'なし',
+    'on': 'オン',
+    'off': 'オフ',
+    'config': 'Config',
+    'enable': 'Enable',
+    'disable': 'Disable',
+    'images': '画像',
+    'oss': 'OSS',
+    'localProvider': 'ローカルprovider',
+    'fullAuto': 'Full auto',
+    'addDirs': '追加ディレクトリ',
+    'skipGitRepoCheck': 'Git repo確認スキップ',
+    'ephemeral': 'Ephemeral',
+    'ignoreUserConfig': 'ユーザー設定を無視',
+    'ignoreRules': 'ルールを無視',
+    'outputSchema': '出力schema',
+    'jsonEvents': 'JSONイベント',
+  },
+  'zh': {
+    'signingIn': '正在登录',
+    'preparingRelay': '正在准备安全中继访问。',
+    'startupFailed': '启动失败',
+    'sessionLoadFailed': '会话加载失败',
+    'commandLoadFailed': '命令加载失败',
+    'newSession': '新建会话',
+    'create': '创建',
+    'creating': '创建中',
+    'connectedAnonymous': '已作为匿名用户连接',
+    'pcBridge': 'PC 桥接',
+    'lastHeartbeat': '上次心跳',
+    'lastQueueCheck': '上次队列检查',
+    'lastManualCheck': '上次手动检查',
+    'lastResponse': '上次响应',
+    'checkPcNow': '立即检查PC',
+    'checking': '检查中',
+    'cliDefaults': 'CLI 默认值',
+    'loading': '加载中',
+    'notifications': '通知',
+    'settings': '设置',
+    'connectionSettings': '连接设置',
+    'notSeenYet': '尚未看到',
+    'sessions': '会话',
+    'sessionCount': '个会话',
+    'loadingSessions': '正在加载会话...',
+    'noSessionsYet': '还没有会话',
+    'noMatchingSessions': '没有匹配的会话',
+    'noCommandsYet': '还没有命令',
+    'searchSessions': '搜索会话',
+    'allGroups': '所有分组',
+    'ungrouped': '未分组',
+    'favorites': '收藏',
+    'favorite': '收藏',
+    'removeFavorite': '取消收藏',
+    'renameSession': '重命名会话',
+    'sessionName': '会话名称',
+    'group': '分组',
+    'groupName': '分组名称',
+    'changeGroup': '更改分组',
+    'deleteSession': '删除会话',
+    'deleteSessionQuestion': '要从应用历史记录中删除此会话吗？',
+    'delete': '删除',
+    'more': '更多',
+    'waitingFinalResult': '正在等待最终结果。',
+    'elapsed': '已用时间',
+    'lastProgress': '上次进度',
+    'instruction': '指令',
+    'send': '发送',
+    'model': '模型',
+    'profile': '配置文件',
+    'sandbox': '沙盒',
+    'bypassSandbox': '绕过沙盒',
+    'bypassSandboxSubtitle': '覆盖沙盒选择',
+    'sandboxUsesDefaults': '沙盒和绕过设置使用CLI默认值。',
+    'advancedCliOptions': '高级CLI选项',
+    'optionalConfigProfile': '可选配置profile',
+    'defaultOption': '默认',
+    'bridgeReadsFinalOutput': '桥接仍从文件读取最终输出',
+    'help': '帮助',
+    'cancel': '取消',
+    'save': '保存',
+    'close': '关闭',
+    'selectImageFile': '选择图片文件',
+    'showHelpFor': '显示帮助',
+    'cliOptionHelp': 'CLI选项帮助',
+    'noHelpAvailable': '此选项没有帮助。',
+    'where': '位置',
+    'example': '示例',
+    'none': '无',
+    'on': '开',
+    'off': '关',
+    'config': '配置',
+    'enable': '启用',
+    'disable': '禁用',
+    'images': '图片',
+    'oss': 'OSS',
+    'localProvider': '本地provider',
+    'fullAuto': '全自动',
+    'addDirs': '追加目录',
+    'skipGitRepoCheck': '跳过Git repo检查',
+    'ephemeral': '临时会话',
+    'ignoreUserConfig': '忽略用户配置',
+    'ignoreRules': '忽略规则',
+    'outputSchema': '输出schema',
+    'jsonEvents': 'JSON事件',
+  },
+  'ko': {
+    'signingIn': '로그인 중',
+    'preparingRelay': '보안 릴레이 접속을 준비 중입니다.',
+    'startupFailed': '시작 실패',
+    'sessionLoadFailed': '세션 로드 실패',
+    'commandLoadFailed': '명령 로드 실패',
+    'newSession': '새 세션',
+    'create': '생성',
+    'creating': '생성 중',
+    'connectedAnonymous': '익명 사용자로 연결됨',
+    'pcBridge': 'PC 브리지',
+    'lastHeartbeat': '마지막 heartbeat',
+    'lastQueueCheck': '마지막 queue 확인',
+    'lastManualCheck': '마지막 수동 확인',
+    'lastResponse': '마지막 응답',
+    'checkPcNow': 'PC 확인',
+    'checking': '확인 중',
+    'cliDefaults': 'CLI 기본값',
+    'loading': '로딩 중',
+    'notifications': '알림',
+    'settings': '설정',
+    'connectionSettings': '연결 설정',
+    'notSeenYet': '아직 확인 안 됨',
+    'sessions': '세션',
+    'sessionCount': '개 세션',
+    'loadingSessions': '세션 로딩 중...',
+    'noSessionsYet': '아직 세션이 없습니다',
+    'noMatchingSessions': '일치하는 세션이 없습니다',
+    'noCommandsYet': '아직 명령이 없습니다',
+    'searchSessions': '세션 검색',
+    'allGroups': '모든 그룹',
+    'ungrouped': '미분류',
+    'favorites': '즐겨찾기',
+    'favorite': '즐겨찾기',
+    'removeFavorite': '즐겨찾기 해제',
+    'renameSession': '세션 이름 변경',
+    'sessionName': '세션 이름',
+    'group': '그룹',
+    'groupName': '그룹 이름',
+    'changeGroup': '그룹 변경',
+    'deleteSession': '세션 삭제',
+    'deleteSessionQuestion': '앱 기록에서 이 세션을 삭제할까요?',
+    'delete': '삭제',
+    'more': '더보기',
+    'waitingFinalResult': '최종 결과를 기다리는 중입니다.',
+    'elapsed': '경과 시간',
+    'lastProgress': '마지막 진행',
+    'instruction': '지시',
+    'send': '전송',
+    'model': '모델',
+    'profile': '프로필',
+    'sandbox': '샌드박스',
+    'bypassSandbox': '샌드박스 우회',
+    'bypassSandboxSubtitle': '샌드박스 선택보다 우선합니다',
+    'sandboxUsesDefaults': '샌드박스와 우회 설정은 CLI 기본값을 사용합니다.',
+    'advancedCliOptions': '고급 CLI 옵션',
+    'optionalConfigProfile': '선택적 config profile',
+    'defaultOption': '기본값',
+    'bridgeReadsFinalOutput': '최종 출력은 계속 파일에서 읽습니다',
+    'help': '도움말',
+    'cancel': '취소',
+    'save': '저장',
+    'close': '닫기',
+    'selectImageFile': '이미지 파일 선택',
+    'showHelpFor': '도움말 표시',
+    'cliOptionHelp': 'CLI 옵션 도움말',
+    'noHelpAvailable': '이 옵션에 대한 도움말이 없습니다.',
+    'where': '위치',
+    'example': '예시',
+    'none': '없음',
+    'on': '켜짐',
+    'off': '꺼짐',
+    'config': '설정',
+    'enable': '활성화',
+    'disable': '비활성화',
+    'images': '이미지',
+    'oss': 'OSS',
+    'localProvider': '로컬 provider',
+    'fullAuto': 'Full auto',
+    'addDirs': '추가 디렉터리',
+    'skipGitRepoCheck': 'Git repo 확인 건너뛰기',
+    'ephemeral': '임시 세션',
+    'ignoreUserConfig': '사용자 설정 무시',
+    'ignoreRules': '규칙 무시',
+    'outputSchema': '출력 schema',
+    'jsonEvents': 'JSON 이벤트',
+  },
+};
